refactor(Widgets): Refactored widgets
