+++ lib/pages/auth_page.dart
@@ -1,7 +1,7 @@
 import 'package:aaveg_app/controllers/nav_bar_controller.dart';
 import 'package:aaveg_app/providers/storage_service.dart';
 import 'package:aaveg_app/views/widgets/NavBar/countdown_timer_widget.dart';
-import 'package:aaveg_app/views/widgets/dauth_web_view.dart';
+import 'package:aaveg_app/views/widgets/AuthWebView/dauth_web_view.dart';
 import 'package:aaveg_app/views/widgets/NavBar/nav_icon_widget.dart';
 import 'package:aaveg_app/views/widgets/NavBar/navbar_widget.dart';
 import 'package:aaveg_app/views/widgets/NavBar/timer_widget.dart';
