+++ lib/pages/my_squad_page.dart
@@ -1,9 +1,9 @@
 import 'package:aaveg_app/controllers/nav_bar_controller.dart';
 import 'package:aaveg_app/views/widgets/NavBar/nav_icon_widget.dart';
 import 'package:aaveg_app/views/widgets/NavBar/navbar_widget.dart';
-import 'package:aaveg_app/views/widgets/my_squad_char_desc.dart';
-import 'package:aaveg_app/views/widgets/my_squad_details.dart';
-import 'package:aaveg_app/views/widgets/my_squad_tile_widget.dart';
+import 'package:aaveg_app/views/widgets/Mysquad/my_squad_char_desc.dart';
+import 'package:aaveg_app/views/widgets/Mysquad/my_squad_details.dart';
+import 'package:aaveg_app/views/widgets/Mysquad/my_squad_tile_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
