+++ lib/views/widgets/Mysquad/my_squad_char_desc.dart
