+++ lib/views/widgets/Mysquad/my_squad_details.dart
