+++ lib/views/widgets/Mysquad/my_squad_tile_widget.dart
@@ -1,4 +1,4 @@
-import 'package:aaveg_app/views/widgets/my_squad_details.dart';
+import 'package:aaveg_app/views/widgets/Mysquad/my_squad_details.dart';
 import 'package:flutter/material.dart';
 
 class MySquadWidget extends StatefulWidget {
